print_inteface_matrix() method seems to work fine. The output is generated in the file interface_matrix_cycle.txt and can be read by matlab.
The print_inteface_matrix() was adapted from the
compute_multiscale_basis() method in the elasticity problem.
But the output is not as expected.

The matrix from stokes problem is not symmetric
ever! But same method in elasticity problem gives
symmetric matrix. So I am not sure if the problem
is with the print_inteface_matrix() method or the
stokes problem itself.