Flag for printing interface matrix. this is a very expensive operation, so use with caution.
Removed bug to run mortar with Q1-discont (RT<dim>(1))
and Q0 (RT<dim>(0)). These can also be plotted but
only on the FE grid, not on the mortar grid. This
is because RTk does not have support points
but we can project on Q2 to plot on the fe grid.

Have to be careful for matlab plot or it will not read the
solution data. Don't ask for plot on mortar grid
if discont mortar space is used (Q1-discont or Q0).